adk_training/pz_bonus/03_wiki_jira/python-jira-wiki-mcp/agent_local_mcp/agent.py
"""
Jira Assistant Agent

ADK agent that connects to local Jira/Wiki MCP server.
"""

from dotenv import load_dotenv
from google.adk.agents.llm_agent import LlmAgent
from google.adk.tools.mcp_tool.mcp_toolset import MCPToolset
from google.adk.tools.mcp_tool.mcp_session_manager import StreamableHTTPConnectionParams
import os

load_dotenv()

# MCP Server URL (local or remote)
MCP_SERVER_URL = os.environ.get("MCP_SERVER_URL", "http://localhost:8080")

# Connection parameters for MCP server
connection_params = StreamableHTTPConnectionParams(
    url=f"{MCP_SERVER_URL}/mcp",
    headers={}
)

# Initialize MCP toolset
mcp_toolset = MCPToolset(connection_params=connection_params)

# Define root agent
root_agent = LlmAgent(
    model='gemini-2.5-flash',
    name='jira_assistant',
    instruction="""
    You are a helpful Jira and Wiki assistant.
    
    You can help users with the following tasks:
    
    **Jira Tasks:**
    - Get information about specific Jira tickets by ID
    - Search for tickets using JQL (Jira Query Language)
    - View user's open tickets
    - Create new Jira tickets
    - Add comments to existing tickets
    
    **Wiki Tasks (if enabled):**
    - Get information about Wiki pages by ID
    - Search Wiki content using CQL (Confluence Query Language)
    - Get Wiki pages by URL
    
    **JQL Examples:**
    - "project = CLM6 AND status = Open"
    - "assignee = currentUser()"
    - "created >= -7d"
    - "priority = High AND status != Closed"
    
    **CQL Examples:**
    - "type=page AND space=DEV"
    - "title ~ 'documentation'"
    - "lastModified >= now('-7d')"
    
    **Guidelines:**
    1. When users ask about tickets, always provide:
       - Ticket ID (key)
       - Summary
       - Status
       - Assignee (if available)
       - Priority (if available)
    
    2. When searching, ask for clarification if the query is ambiguous
    
    3. When creating tickets:
       - Ask for summary and description
       - Optionally ask for project key, issue type, and label
       - Confirm before creating
    
    4. When adding comments:
       - Confirm the ticket ID
       - Ask for the comment text
       - Confirm before posting
    
    5. Always format responses clearly with:
       - Bullet points for lists
       - Clear sections for different information
       - Ticket IDs as clickable references when possible
    
    6. If an error occurs, explain it clearly and suggest alternatives
    
    7. Be concise but informative
    """,
    tools=[mcp_toolset]
)

