adk_training/pz_bonus/03_wiki_jira/python-jira-wiki-mcp/requirements.txt
# Jira/Wiki MCP Server Requirements

# FastMCP - MCP server framework
fastmcp

# HTTP requests
requests

# Environment variables
python-dotenv

# Google ADK (for agent)
google-adk

